+++ lib/src/service_locator_adapter.dart
@@ -9,29 +9,37 @@ abstract class ServiceLocatorAdapter {
     required String? name,
     required dynamic key,
     required String? environment,
+    required bool allowMultipleInstances,
   });
 
   /// Registers a factory for type [T].
   Triplet<Success, Conflict, InternalError> registerFactory<T extends Object>(
-    T Function(ServiceLocator serviceLocator) factory, {
+    T Function(
+      ServiceLocator serviceLocator,
+      Map<String, dynamic> namedArgs,
+    ) factory, {
     required List<Type>? interfaces,
     required String? name,
     required dynamic key,
     required String? environment,
+    required bool allowMultipleInstances,
   });
 
   /// Retrieves an instance of type [T].
-  Triplet<T, NotFound, InternalError> getFirst<T extends Object>({
+  Triplet<T, NotFound, InternalError> resolve<T extends Object>({
     required String? name,
     required dynamic key,
     required String? environment,
+    required Map<String, dynamic> namedArgs,
+    required bool resolveFirst,
   });
 
   /// Retrieves a List of instances of type [T].
-  Triplet<List<T>, NotFound, InternalError> getAll<T extends Object>({
+  Triplet<List<T>, NotFound, InternalError> resolveAll<T extends Object>({
     required String? name,
     required dynamic key,
     required String? environment,
+    required Map<String, dynamic> namedArgs,
   });
 
   /// Overrides an existing registration with a new instance of type [T].
@@ -44,7 +52,10 @@ abstract class ServiceLocatorAdapter {
 
   /// Overrides an existing registration with a new factory of type [T].
   Doublet<Success, InternalError> overrideFactory<T extends Object>(
-    T Function(ServiceLocator serviceLocator) factory, {
+    T Function(
+      ServiceLocator serviceLocator,
+      Map<String, dynamic> namedArgs,
+    ) factory, {
     required String? name,
     required dynamic key,
     required String? environment,