+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0
+
+- Possibility to pass arguments to factory resolutions
+- replaced `getFirst()` with `resolve()`
+- replaced `getFirstOrNull()` with `resolveOrNull()`
+- replaced `getAll()` with `resolveAll()`
+
 ## 0.0.9
 
 - Implemented usage of zef_log_core